LayoutTests/fast/images/text-recognition/mac/context-menu-for-image-in-link-contains-copy-expected.txt

PASS getCopyContextMenuItem() is null
PASS Selected text in image
PASS getCopyContextMenuItem() is non-null.
PASS successfullyParsed is true

TEST COMPLETE
